+++ pycvvdp/utils.py
@@ -149,7 +149,7 @@ class config_files:
             if not (os.path.isfile(cp) or os.path.isdir(cp)):
                 raise RuntimeError( f"config_path '{cp}' does not exist" )
 
-            if os.path.isfile(cp) and os.path.basename(cp).startswith(bname):
+            if os.path.isfile(cp) and os.path.basename(cp).startswith(bname) and os.path.basename(cp).endswith(ext):
                 return cp
 
         # Then, check all directories